+++ zebra/subdir.am
@@ -120,6 +120,7 @@ zebra_zebra_SOURCES = \
 	zebra/zebra_vrf.c \
 	zebra/zebra_vty.c \
 	zebra/zebra_vxlan.c \
+	zebra/zebra_vxlan_if.c \
 	zebra/zebra_evpn_mh.c \
 	zebra/zserv.c \
 	# end
